share oteltest setup with golden tests
